+++ src/main/java/com/breakinblocks/plonk/common/registry/RegistryPackets.java
@@ -2,6 +2,7 @@ package com.breakinblocks.plonk.common.registry;
 
 import com.breakinblocks.plonk.Plonk;
 import com.breakinblocks.plonk.common.packet.PacketPlaceItem;
+import com.breakinblocks.plonk.common.packet.PacketRotateTile;
 import net.minecraftforge.fml.common.network.simpleimpl.SimpleNetworkWrapper;
 import net.minecraftforge.fml.relauncher.Side;
 
@@ -12,5 +13,6 @@ public class RegistryPackets {
         int id = 0;
         // Client -> Server
         CHANNEL.registerMessage(PacketPlaceItem.class, PacketPlaceItem.class, id++, Side.SERVER);
+        CHANNEL.registerMessage(PacketRotateTile.class, PacketRotateTile.class, id++, Side.SERVER);
     }
 }